move document subscription to different domain
